requirements.txt
django
graphene-django
hvac
python-consul